Add requirements.in for specifying looser dependencies.
The pip-tools package provides a pip-compile command to convert
the requirements.in file into a requirements.txt file.

This lets us specify the desired constraints on dependencies in the `in`
file, and makes it easier to do updates and record all the exact
transitive dependency versions in requirements.txt when we do updates.